moved forRoot initialization for Localize-router and translate modules to app.module
