+++ src/app/shared/shared.module.ts
@@ -1,24 +1,13 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { Http, HttpModule } from '@angular/http';
-import { HttpClient, HttpClientModule } from '@angular/common/http';
-import { Location } from '@angular/common';
+import { HttpModule } from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 
-import { LocalizeParser, LocalizeRouterModule, StaticParserLoader } from 'localize-router';
-import { TranslateModule, TranslateLoader, TranslateService } from '@ngx-translate/core';
-import { TranslateHttpLoader } from '@ngx-translate/http-loader';
+import { LocalizeRouterModule } from 'localize-router';
+import { TranslateModule } from '@ngx-translate/core';
 
 import { MaterialModule } from './material.module';
-
-// AoT requires an exported function for factories.
-export function HttpLoaderFactory(http: HttpClient) {
-	return new TranslateHttpLoader(http, 'assets/i18n/', '.json');
-}
-
-export function localizeLoaderFactory(translate: TranslateService, location: Location, http: Http) {
-	return new StaticParserLoader(translate, location, http, 'assets/i18n/locales.json');
-}
 
 @NgModule({
 	imports     : [
@@ -27,18 +16,8 @@ export function localizeLoaderFactory(translate: TranslateService, location: Loc
 		HttpModule,
 		HttpClientModule,
 		MaterialModule,
-		LocalizeRouterModule.forRoot([], {
-			provide: LocalizeParser,
-			useFactory: localizeLoaderFactory,
-			deps: [TranslateService, Location, Http],
-		}),
-		TranslateModule.forRoot({
-			loader: {
-				provide: TranslateLoader,
-				useFactory: HttpLoaderFactory,
-				deps: [ HttpClient ],
-			},
-		}),
+		LocalizeRouterModule,
+		TranslateModule,
 	],
 	declarations: [],
 	providers   : [],
